box logic
I changed the "object" that can be picked up to be a "box", we can change it later.

box can now drop/pick up when moving horizontally.

box can now "stick" on moving platforms, as well as stacking on moving platforms.

player can "stick" to boxes on moving platforms.

logic is a little wonky and hard to read but idk